+++ src/Nuprl/Tactic.hs
@@ -38,6 +38,7 @@ module Nuprl.Tactic
   , tacUnfold
   , tacReduce
   , tacDecideInt
+  , tacDecideLt
   , tacCases
     -- * Automation
   , tacAuto
@@ -228,6 +229,10 @@ tacExact t = tacCut t (Var "h") `tacThen` tacIntro (IntroWitness t) `tacThen` ta
 tacDecideInt :: Term -> Term -> Tactic
 tacDecideInt a b = primTac "decide" (RuleDecideInt a b)
 
+-- | Case analysis on integer comparison.
+tacDecideLt :: Term -> Term -> Tactic
+tacDecideLt a b = primTac "decide" (RuleDecideLt a b)
+
 -- | Case analysis on a term of union type.
 tacCases :: Term -> Tactic
 tacCases t = primTac "decide" (RuleCases t)
@@ -356,6 +361,7 @@ data TacticExpr
   | TxLemma Name [Term]
   | TxThin Int
   | TxDecideInt Term Term
+  | TxDecideLt Term Term
   | TxCases Term
   | TxThen TacticExpr TacticExpr
   | TxThenL TacticExpr [TacticExpr]
@@ -404,6 +410,7 @@ evalTactic = \case
   TxLemma n args -> tacLemma n args
   TxThin i -> tacThin i
   TxDecideInt a b -> tacDecideInt a b
+  TxDecideLt a b -> tacDecideLt a b
   TxCases t -> tacCases t
   TxThen a b -> evalTactic a `tacThen` evalTactic b
   TxThenL a bs -> tacThenL (evalTactic a) (map evalTactic bs)
@@ -446,6 +453,7 @@ prettyTacticExpr = \case
   TxLemma n _ -> "lemma " <> n
   TxThin i -> "thin " <> tshow i
   TxDecideInt {} -> "decide … = …"
+  TxDecideLt {} -> "decide … < …"
   TxCases {} -> "decide …"
   TxThen a b -> prettyTacticExpr a <> " THEN " <> prettyTacticExpr b
   TxThenL a _ -> prettyTacticExpr a <> " THENL …"